feat: filter, tambah penerbangan, dan html

+++ sipes/src/main/java/apap/tugas/sipes/controller/PesawatController.java
@@ -15,6 +15,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
+
 import java.util.*;
 
 @Controller
@@ -51,14 +52,82 @@ public class PesawatController {
             @PathVariable Long id,
             Model model
     ) {
+        // ambil pesawat dengan id
         PesawatModel pesawat = pesawatService.getPesawatById(id);
 
         // list teknisi di pesawat
         List<TeknisiModel> listTeknisi = pesawat.getListTeknisi();
+
+        // list seluruh penerbangan
+        List<PenerbanganModel> listAllPenerbangan = penerbanganService.getListPenerbangan();
+
+        // list penerbangan di pesawat saat ini
+        List<PenerbanganModel> listPenerbangan = pesawat.getListPenerbangan();
+
+        // mengambil penerbangan yang belum di assign
+        List<PenerbanganModel> pilihan = new ArrayList<>();
+        for(PenerbanganModel penerbangan : listAllPenerbangan){
+            if(penerbangan.getPesawat() != pesawat){
+                pilihan.add(penerbangan);
+            }
+        }
+
         model.addAttribute("pesawat", pesawat);
         model.addAttribute("listTeknisi", listTeknisi);
+        model.addAttribute("pilihan", pilihan);
+        model.addAttribute("listPenerbangan", listPenerbangan);
         return "view-detail-pesawat";
     }
+
+
+    @PostMapping("/pesawat/{id}/tambah-penerbangan")
+    public String addPenerbanganToPesawat(
+            @PathVariable Long id,
+            @RequestParam(value="idSelected") Long idSelected,
+            Model model
+    ){
+        // ambil pesawat dengan id
+        PesawatModel pesawat = pesawatService.getPesawatById(id);
+
+        // list teknisi di pesawat
+        List<TeknisiModel> listTeknisi = pesawat.getListTeknisi();
+
+        // list seluruh penerbangan
+        List<PenerbanganModel> listAllPenerbangan = penerbanganService.getListPenerbangan();
+
+        // list penerbangan di pesawat saat ini
+        List<PenerbanganModel> listPenerbangan = pesawat.getListPenerbangan();
+
+        // mengambil penerbangan yang belum di assign
+        List<PenerbanganModel> pilihan = new ArrayList<>();
+        for(PenerbanganModel penerbangan : listAllPenerbangan){
+            if(penerbangan.getPesawat() != pesawat){
+                pilihan.add(penerbangan);
+            }
+        }
+
+        // mengambil penerbangan berdasarkan id penerbangan yang telah dipilih
+        PenerbanganModel penerbangan = penerbanganService.getPenerbanganById(idSelected);
+
+        // menambahkan penerbangan ke pesawat lalu update
+        pesawat.getListPenerbangan().add(penerbangan);
+        pesawatService.updatePesawat(pesawat);
+        // menambahkan pesawat ke penerbangan lalu update
+        penerbangan.setPesawat(pesawat);
+        penerbanganService.updatePenerbangan(penerbangan);
+
+        // membuat notifikasi, mengambil penerbangan terakhir
+        PenerbanganModel temp = listPenerbangan.get(listPenerbangan.size()-1);
+        String notifikasi = "Penerbangan dengan nomor "+temp.getNomor_penerbangan()+" berhasil ditambahkan";
+
+        model.addAttribute("pesawat", pesawat);
+        model.addAttribute("listTeknisi", listTeknisi);
+        model.addAttribute("pilihan", pilihan);
+        model.addAttribute("listPenerbangan", listPenerbangan);
+        model.addAttribute("notifikasi", notifikasi);
+        return "view-detail-pesawat";
+    }
+
 
     @GetMapping("/pesawat/tambah")
     public String addPesawatForm(
@@ -88,7 +157,6 @@ public class PesawatController {
         model.addAttribute("listAllTeknisi", listAllTeknisi);
         return "form-add-pesawat";
     }
-
 
     @PostMapping(value = "/pesawat/tambah", params = {"tambah"})
     public String addTeknisi(
@@ -158,6 +226,13 @@ public class PesawatController {
     ) {
         // ambil pesawat dari idnya
         PesawatModel pesawat = pesawatService.getPesawatById(id);
+
+        // kalau punya penerbangan gaboleh diapus
+        if (pesawat.getListPenerbangan().size() > 0){
+            model.addAttribute("pesawat", pesawat);
+            return "cannot-delete";
+        }
+        else
         // delete pesawat di service
         pesawatService.deletePesawat(pesawat);
         model.addAttribute("pesawat", pesawat);
@@ -205,5 +280,78 @@ public class PesawatController {
         model.addAttribute("pesawat", pesawat);
         model.addAttribute("total", total);
         return "bonus-total-teknisi";
+    }
+
+    @GetMapping("/pesawat/filter")
+    private String filterPesawat(
+            @RequestParam(value= "idPenerbangan", defaultValue = "0") Long idPenerbangan,
+            @RequestParam(value= "idTipe", defaultValue = "0") Long idTipe,
+            @RequestParam(value= "idTeknisi", defaultValue = "0") Long idTeknisi,
+            Model model
+    ) {
+        // ambil list tipe, teknisi, dan penerbangan
+        List<TipeModel> listTipe = tipeService.getListTipe();
+        List<TeknisiModel> listTeknisi = teknisiService.getListTeknisi();
+        List<PenerbanganModel> listPenerbangan = penerbanganService.getListPenerbangan();
+
+        // tampilin
+        model.addAttribute("listTeknisi", listTeknisi);
+        model.addAttribute("listTipe", listTipe);
+        model.addAttribute("listPenerbangan", listPenerbangan);
+
+        // kondisi semuanya tidak null
+        Boolean test = (idPenerbangan != null || idTipe != null || idTeknisi != null);
+        if (test == Boolean.TRUE) {
+            // ambil list pesawat
+            List<PesawatModel> listPesawat = pesawatService.getPesawatList();
+
+            // bikin list buat nyimpen
+            List<PesawatModel> simpenPesawat = new ArrayList<PesawatModel>();
+
+            // masukin pesawat ke list
+            for (PesawatModel pesawat : listPesawat) {
+                simpenPesawat.add(pesawat);
+            }
+
+            for (PesawatModel pesawat : simpenPesawat) {
+                // filter untuk penerbangan
+                if (idPenerbangan != 0) {
+                    // cari penerbangan dari idnya
+                    PenerbanganModel penerbangan = penerbanganService.getPenerbanganById(idPenerbangan);
+                    // ambil list penerbangan
+                    List<PenerbanganModel> listPen = pesawat.getListPenerbangan();
+                    // menghapus penerbangan yang tidak di-get
+                    if (!listPen.contains(penerbangan)) {
+                        listPesawat.remove(pesawat);
+                    }
+                }
+                // filter untuk teknisi
+                if (idTeknisi != 0) {
+                    // cari teknisi dari idnya
+                    TeknisiModel teknisi = teknisiService.getTeknisiById(idTeknisi);
+                    // ambil list teknisi
+                    List<TeknisiModel> listTek = pesawat.getListTeknisi();
+                    // menghapus teknisi yang tidak di-get
+                    if (!listTek.contains(teknisi)) {
+                        listPesawat.remove(pesawat);
+                    }
+                }
+                // filter untuk tipe
+                if (idTipe != 0) {
+                    // cari tipe dari idnya
+                    TipeModel tipe = tipeService.getTipeById(idTipe);
+                    // ambil tipe pesawat
+                    TipeModel tipePes = pesawat.getTipe();
+                    // menghapus tipe yang tidak sama dengan yang di-get
+                    if (tipe.getId() != tipePes.getId()) {
+                        listPesawat.remove(pesawat);
+                    }
+                }
+            }
+            Integer size = listPesawat.size();
+            model.addAttribute("listPesawat", listPesawat);
+            model.addAttribute("size", size);
+        }
+        return "pesawat-filter";
     }
 }
